Add another VAO and window resize callback
Here I was trying to understand more about VAO. To do that, I was testing some things to see how VAO behaves. It did answer some of my curiosity too. Anyway, this is something I made while messing around with it. Man, Finding the momentum to make it like this was kinda a pain haha.